Restore NinSnes to master behavior
