+++ src/main/CMakeLists.txt
@@ -114,6 +114,7 @@ target_sources(vgmtranscore
     components/VGMFile.cpp
     components/VGMItem.cpp
     components/VGMMiscFile.cpp
+    components/VGMMultiSectionSeq.cpp
     components/VGMSampColl.cpp
     components/VGMTag.cpp
         components/instr/Modulation.cpp
@@ -128,6 +129,7 @@ target_sources(vgmtranscore
     components/seq/SeqTrack.cpp
     components/seq/VGMSeq.cpp
     components/seq/VGMSeqNoTrks.cpp
+    components/seq/VGMSeqSection.cpp
     conversion/DLSConversion.cpp
     conversion/DLSFile.cpp
     conversion/MidiFile.cpp
@@ -225,8 +227,6 @@ target_sources(vgmtranscore
     formats/NinSnes/NinSnesScannerPatterns.cpp
     formats/NinSnes/NinSnesSeq.cpp
     formats/NinSnes/NinSnesSeqIntelli.cpp
-    formats/NinSnes/NinSnesTrack.cpp
-    formats/NinSnes/NinSnesTrackModulation.cpp
     formats/Org/OrgScanner.cpp
     formats/Org/OrgSeq.cpp
     formats/PS1/PS1Seq.cpp
@@ -483,6 +483,7 @@ target_sources(vgmtranscore
       components/VGMFile.h
       components/VGMItem.h
       components/VGMMiscFile.h
+      components/VGMMultiSectionSeq.h
       components/VGMSampColl.h
       components/VGMTag.h
     FILE_SET headers_components_instr TYPE HEADERS BASE_DIRS components/instr
@@ -509,6 +510,7 @@ target_sources(vgmtranscore
       components/seq/SeqTrack.h
       components/seq/VGMSeq.h
       components/seq/VGMSeqNoTrks.h
+      components/seq/VGMSeqSection.h
     FILE_SET headers_conversion TYPE HEADERS BASE_DIRS conversion
     FILES
       conversion/DLSConversion.h
